pyproject.toml
[project]
name = "discord-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
